Moves NavigationViewControllerDelegate extension to POIInfoViewController+Navigation+Delegate file
