+++ quavi/quavi/Controllers/POI Info/POIInfoViewController+Navigation+Delegate.swift
@@ -7,3 +7,21 @@
 //
 
 import Foundation
+import MapboxNavigation
+import MapboxDirections
+import MapboxCoreNavigation
+
+extension POIInfoViewController: NavigationViewControllerDelegate{
+    func navigationViewController(_ navigationViewController: NavigationViewController, didArriveAt waypoint: Waypoint) -> Bool {
+        let waypointCount = (selectedRoute?.routeOptions.waypoints.count)!
+         if nextStopIndex == waypointCount - 1{
+                isAtLastLeg = true
+                }else{
+                isAtLastLeg = false
+                }
+        navigationViewController.navigationService.stop()
+        navigationViewController.navigationService.endNavigation(feedback: nil)
+        navigationViewController.dismiss(animated: true, completion: nil)
+        return false
+    }
+}
