+++ quavi/quavi/Controllers/POI Info/POIInfoViewController+objc.swift
@@ -95,17 +95,3 @@ extension POIInfoViewController {
 
 }
 
-extension POIInfoViewController: NavigationViewControllerDelegate{
-    func navigationViewController(_ navigationViewController: NavigationViewController, didArriveAt waypoint: Waypoint) -> Bool {
-        let waypointCount = (selectedRoute?.routeOptions.waypoints.count)!
-         if nextStopIndex == waypointCount - 1{
-                isAtLastLeg = true
-                }else{
-                isAtLastLeg = false
-                }
-        navigationViewController.navigationService.stop()
-        navigationViewController.navigationService.endNavigation(feedback: nil)
-        navigationViewController.dismiss(animated: true, completion: nil)
-        return false
-    }
-}
